Level editor rework and main menu select
Reworked the level editor as I couldn't get it to function the way I intended. Data wasn't saving the the scriptable object and I'm not really sure why.
Levels can be generated with the grid and tile size and positioned to where they fit the screen. Tiles then need to be added into the list in the inspector of the LevelDataSO

Main menu added with a play button, level select and a quit button. Level select loads the levels based on it's index in the GameManagers level array. Perhaps could create an enum for readability. Levels load nicely though and have relevant ui
Need to add button to open menu during gameplay